Proof of concept - cancellation of https.get
NOTE: This changes the behaviour of http.ClientRequest! In order to be
consistent with the other cancellation features in this branch, we emit
"error" with ERR_CANCELLED as well as emitting "abort" when the request
is aborted.

NOTE: Occasionally the main loop will wait for 1 or 2 seconds after the
request is cancelled. Need to investigate this still.